DIY Weather_Station using Nodemcu, BMP180, DHT11 with Arduino IDE over Blynk Server
This is a DIY Weather Station which uses BMP180 and DHT11. BMP180 gives the Temperature, Pressure and the Altitude value of the current place. DHT11 gives the Temerature and Humidity as the output. The data from the both sensors is uploaded to the Blynk Server and can be viewd using the Blynk App.